Regresión Lineal/educacion/requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.13
scipy>=1.10
scikit-learn>=1.3
statsmodels>=0.14
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
jupyter>=1.0
